refactor(palette): split db/src/palette.rs into palette/ submodules
Splits by query family (fts5 / overlay / ranking) per file-shape rule.
Each submodule has a short //!. No behavior changes.
PR1 of stylistic cleanup — sqlx::Error wrapping deferred to #305.

Closes #309
